The Java Files Containing the Diet Plan Program
These are the 4 classes in the Diet Plan program I created for myself. The Food class contains the constructor of the Food object, a way to print the stats of what I've eaten today, and an AddFood method used to add a food to what I've eaten today. The Day class defines what a Day object is that will hold the macronutrient information I've eaten today, as well as a printStats method that prints out 7 days worth of information and whether or not I've hit my goals. The Week class defines the Week object that holds 7 days worth of data in it. The FoodObjects method is the main method holding the different Food objects I've input as well as the calls to the various methods produced in the other classes.